Shopping_Assistant/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
tavily-python>=0.3.0
pydantic>=2.0.0
streamlit>=1.38.0
streamlit-extras>=0.4.0
python-dotenv>=1.0.0
requests>=2.31.0
diskcache>=5.6.0
